refactor: 기존 OAuth 관련 클래스 제거 (#98)
