--- src/main/java/com/jisungin/api/oauth/AuthConstant.java
@@ -1,7 +1,0 @@
-package com.jisungin.api.oauth;
-
-public class AuthConstant {
-
-    public static final String JSESSION_ID = "JSESSIONID";
-
-}
