--- src/main/java/com/jisungin/api/oauth/AuthContext.java
@@ -1,20 +1,0 @@
-package com.jisungin.api.oauth;
-
-import org.springframework.stereotype.Component;
-import org.springframework.web.context.annotation.RequestScope;
-
-@Component
-@RequestScope
-public class AuthContext {
-
-    private Long userId;
-
-    public void setUserId(Long userId) {
-        this.userId = userId;
-    }
-
-    public Long getUserId() {
-        return userId;
-    }
-
-}
