--- src/main/java/com/jisungin/domain/oauth/client/UserClient.java
@@ -1,14 +1,0 @@
-package com.jisungin.domain.oauth.client;
-
-import com.jisungin.domain.oauth.OauthType;
-import com.jisungin.domain.user.User;
-
-public interface UserClient {
-
-    OauthType supportType();
-
-    User fetch(String authCode);
-
-    void logout(String oauthId);
-
-}
