--- src/main/java/com/jisungin/infra/oauth/authcode/KakaoAuthCodeRequestUrlProvider.java
@@ -1,31 +1,0 @@
-package com.jisungin.infra.oauth.authcode;
-
-import com.jisungin.domain.oauth.OauthType;
-import com.jisungin.domain.oauth.authcode.AuthCodeRequestUrlProvider;
-import com.jisungin.infra.oauth.kakao.KakaoOauthConfig;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-import org.springframework.web.util.UriComponentsBuilder;
-
-@Component
-@RequiredArgsConstructor
-public class KakaoAuthCodeRequestUrlProvider implements AuthCodeRequestUrlProvider {
-
-    private final KakaoOauthConfig kakaoOauthConfig;
-
-    @Override
-    public OauthType supportType() {
-        return OauthType.KAKAO;
-    }
-
-    @Override
-    public String provide() {
-        return UriComponentsBuilder
-                .fromUriString("https://kauth.kakao.com/oauth/authorize")
-                .queryParam("client_id", kakaoOauthConfig.clientId())
-                .queryParam("redirect_uri", kakaoOauthConfig.redirectUri())
-                .queryParam("response_type", "code")
-                .queryParam("scope", String.join(",", kakaoOauthConfig.scope()))
-                .toUriString();
-    }
-}
