--- src/main/java/com/jisungin/infra/oauth/config/HttpInterfaceConfig.java
@@ -1,25 +1,0 @@
-package com.jisungin.infra.oauth.config;
-
-import com.jisungin.infra.oauth.kakao.client.KakaoApiClient;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.reactive.function.client.WebClient;
-import org.springframework.web.reactive.function.client.support.WebClientAdapter;
-import org.springframework.web.service.invoker.HttpServiceProxyFactory;
-
-@Configuration
-public class HttpInterfaceConfig {
-
-    @Bean
-    public KakaoApiClient kakaoApiClient() {
-        return createHttpInterface(KakaoApiClient.class);
-    }
-
-    private <T> T createHttpInterface(Class<T> tClass) {
-        WebClient webClient = WebClient.create();
-        HttpServiceProxyFactory build = HttpServiceProxyFactory
-                .builder(WebClientAdapter.forClient(webClient)).build();
-        return build.createClient(tClass);
-    }
-
-}
