--- src/main/java/com/jisungin/infra/oauth/kakao/KakaoOauthConfig.java
@@ -1,13 +1,0 @@
-package com.jisungin.infra.oauth.kakao;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "oauth.kakao")
-public record KakaoOauthConfig(
-        String clientId,
-        String clientSecret,
-        String redirectUri,
-        String[] scope,
-        String adminKey
-) {
-}
