--- src/main/java/com/jisungin/infra/oauth/kakao/dto/KakaoToken.java
@@ -1,16 +1,0 @@
-package com.jisungin.infra.oauth.kakao.dto;
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-
-@JsonNaming(SnakeCaseStrategy.class)
-public record KakaoToken(
-        String tokenType,
-        String accessToken,
-        String idToken,
-        Integer expiresIn,
-        String refreshToken,
-        Integer refreshTokenExpiresIn,
-        String scope
-) {
-}
