--- src/main/java/com/jisungin/infra/oauth/kakao/dto/KakaoUserResponse.java
@@ -1,70 +1,0 @@
-package com.jisungin.infra.oauth.kakao.dto;
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import com.jisungin.domain.oauth.OauthId;
-import com.jisungin.domain.user.User;
-
-import java.time.LocalDateTime;
-
-import static com.jisungin.domain.oauth.OauthType.*;
-
-@JsonNaming(SnakeCaseStrategy.class)
-public record KakaoUserResponse(
-        Long id,
-        boolean hasSignedUp,
-        LocalDateTime connectedAt,
-        KakaoAccount kakaoAccount
-) {
-
-    public User toEntity() {
-        return User.builder()
-                .oauthId(OauthId.builder()
-                        .oauthId(String.valueOf(id))
-                        .oauthType(KAKAO)
-                        .build()
-                )
-                .name(kakaoAccount.profile.nickname)
-                .profileImage(kakaoAccount.profile.profileImageUrl)
-                .build();
-    }
-
-    @JsonNaming(SnakeCaseStrategy.class)
-    public record KakaoAccount(
-            boolean profileNeedsAgreement,
-            boolean profileNicknameNeedsAgreement,
-            boolean profileImageNeedsAgreement,
-            Profile profile,
-            boolean nameNeedsAgreement,
-            String name,
-            boolean emailNeedsAgreement,
-            boolean isEmailValid,
-            boolean isEmailVerified,
-            String email,
-            boolean ageRangeNeedsAgreement,
-            String ageRange,
-            boolean birthyearNeedsAgreement,
-            String birthyear,
-            boolean birthdayNeedsAgreement,
-            String birthday,
-            String birthdayType,
-            boolean genderNeedsAgreement,
-            String gender,
-            boolean phoneNumberNeedsAgreement,
-            String phoneNumber,
-            boolean ciNeedsAgreement,
-            String ci,
-            LocalDateTime ciAuthenticatedAt
-    ) {
-    }
-
-    @JsonNaming(SnakeCaseStrategy.class)
-    public record Profile(
-            String nickname,
-            String thumbnailImageUrl,
-            String profileImageUrl,
-            boolean isDefaultImage
-    ) {
-    }
-
-}
