+++ src/test/java/com/jisungin/application/rating/RatingServiceTest.java
@@ -4,10 +4,10 @@ import com.jisungin.ServiceTestSupport;
 import com.jisungin.application.rating.request.RatingCreateServiceRequest;
 import com.jisungin.domain.book.Book;
 import com.jisungin.domain.book.repository.BookRepository;
-import com.jisungin.domain.oauth.OauthId;
-import com.jisungin.domain.oauth.OauthType;
 import com.jisungin.domain.rating.Rating;
 import com.jisungin.domain.rating.repository.RatingRepository;
+import com.jisungin.domain.user.OauthId;
+import com.jisungin.domain.user.OauthType;
 import com.jisungin.domain.user.User;
 import com.jisungin.domain.user.repository.UserRepository;
 import org.junit.jupiter.api.AfterEach;
