+++ src/test/java/com/jisungin/domain/rating/repository/RatingRepositoryImplTest.java
@@ -5,9 +5,9 @@ import com.jisungin.application.PageResponse;
 import com.jisungin.application.rating.response.RatingGetResponse;
 import com.jisungin.domain.book.Book;
 import com.jisungin.domain.book.repository.BookRepository;
-import com.jisungin.domain.oauth.OauthId;
-import com.jisungin.domain.oauth.OauthType;
 import com.jisungin.domain.rating.Rating;
+import com.jisungin.domain.user.OauthId;
+import com.jisungin.domain.user.OauthType;
 import com.jisungin.domain.user.User;
 import com.jisungin.domain.user.repository.UserRepository;
 import org.junit.jupiter.api.AfterEach;
